chore: api 폴더 중복 해결 (#151)
