--- src/api/endpoints.ts
@@ -1,6 +1,0 @@
-export const API_ENDPOINTS = {
-  SIGNUP: '/users/signup',
-  LOGIN: 'login',
-  LOGOUT: '/logout',
-  REFRESH_TOKEN: '/token/refresh',
-};
