--- src/api/index.ts
@@ -1,3 +1,0 @@
-export * from './authAPI';
-export * from './backendAPI';
-export * from './endpoints';
