+++ src/features/auth/api/authAPI.ts
@@ -1,10 +1,12 @@
 import { API_ENDPOINTS, backendAPI } from '@/features/auth/api';
 
-export const authAPI = {
-  login: (data: { email: string; password: string }) => backendAPI.post(API_ENDPOINTS.LOGIN, data),
-  signup: (data: { email: string; password: string }) =>
-    backendAPI.post(API_ENDPOINTS.SIGNUP, data),
-  logout: () => backendAPI.post(API_ENDPOINTS.LOGOUT),
-  refreshToken: (data: { refreshToken: string }) =>
-    backendAPI.post(API_ENDPOINTS.REFRESH_TOKEN, data),
-};
+export const authLogin = (data: { email: string; password: string }) =>
+  backendAPI.post(API_ENDPOINTS.LOGIN, data);
+
+export const authSignup = (data: { email: string; password: string }) =>
+  backendAPI.post(API_ENDPOINTS.SIGNUP, data);
+
+export const authLogout = () => backendAPI.post(API_ENDPOINTS.LOGOUT);
+
+export const authRefreshToken = (data: { refreshToken: string }) =>
+  backendAPI.post(API_ENDPOINTS.REFRESH_TOKEN, data);
