+++ src/features/auth/index.ts
@@ -1,3 +1,4 @@
 export * from './GlobalModalManager';
+export * from './api';
 export * from './components';
 export * from './store';
